removed unnecessary type check of filterHori/Verti
The purpose of this check is to ensure that the value of filterHori is a number before using it to filter the displays array. Without this check, if filterHori were to contain a non-number value (such as null or undefined), the displays.filter method would throw an error when it tries to compare the horizontal_resolution property of each DisplayData object to that value.

However, since filterHori is only being used in a context where it is guaranteed to be a number (i.e., in the Input component where onInput is setting it to parseInt(target.value)), the typeof filterHori() === "number" check may be redundant.

That being said, including the check is a defensive programming practice that can help prevent unexpected errors in case the implementation of the Input component were to change in the future.